Extract checkStatus into lib/cluster/status.js
